Extract PYNQ driver as standalone finn-plus-driver subpackage
Refactor the PYNQ runtime driver from an internal compiler template into
a PyPI-publishable subpackage (finn_plus_driver) that:

- Splits monolithic driver.py (1541 lines, 4 overlay classes + CLI) into
  per-class modules under overlays/ for improved maintainability
- Vendors minimal packing.py runtime helpers to avoid onnx/brevitas deps
  on ARM boards
- Moves dataset validation code from templates/validate/ into driver
  with lazy-import dispatch (no sys.path hacks)
- Extracts CLI as a console script entry point via __main__.py
- Simplifies MakePYNQDriver to emit only settings.json + runtime_weights,
  no source copying
- Updates board CI (driver_install_dependencies.sh) to pip install from
  checkout
- Configures GitHub release workflow to publish driver alongside tests
  and notebooks

The driver is now installable as: pip install ./driver (dev) or from
PyPI once released. Board images continue to run the same overlay classes,
just now as an imported package rather than copied template source.

Core changes:
- New: driver/ subpackage with pyproject.toml, README, overlays, packing,
  validate, cli
- Moved: templates/python_driver/driver.py -> overlays/
- Moved: templates/validate/* -> driver/validate/
- Simplified: make_driver.py now skips source copy
- Updated: experiment_manager.py uses finn_plus_driver imports
- Updated: .github/workflows/release.yml publishes driver
- Cleaned: Removed obsolete driver path from experiment configs
