+++ driver/finn_plus_driver/__init__.py
@@ -1,0 +1,48 @@
+"""Runtime driver for FPGA accelerators generated by finn-plus.
+
+The overlay classes require ``pynq``, which is only available on a PYNQ board image. They
+are therefore imported lazily, so that helper modules such as :mod:`finn_plus_driver.packing`
+remain usable off-board (for example in tests or on a build host).
+"""
+
+from typing import Any
+
+#: Maps the driver type names emitted into ``settings.json`` by the finn-plus build flow to
+#: the module that implements them. Used instead of module introspection so the mapping is
+#: explicit and does not depend on import side effects.
+DRIVER_TYPES: dict[str, str] = {
+    "FINNDMAOverlay": "finn_plus_driver.overlays.dma",
+    "FINNInstrumentationOverlay": "finn_plus_driver.overlays.instrumentation",
+    "FINNLiveFIFOOverlay": "finn_plus_driver.overlays.live_fifo",
+    "FINNDMAInstrumentationOverlay": "finn_plus_driver.overlays.dma_instrumentation",
+}
+
+__all__ = [*DRIVER_TYPES, "DRIVER_TYPES", "get_driver_class"]
+
+
+def get_driver_class(driver_type: str) -> type:
+    """Return the overlay class for the given driver type name.
+
+    Args:
+        driver_type: Name of the driver class, as stored in ``settings.json``.
+
+    Returns:
+        The overlay class matching ``driver_type``.
+
+    Raises:
+        ValueError: If ``driver_type`` is not a known driver type.
+    """
+    from importlib import import_module
+
+    if driver_type not in DRIVER_TYPES:
+        raise ValueError(
+            f"Unknown driver type '{driver_type}'. Known types: {sorted(DRIVER_TYPES)}"
+        )
+    return getattr(import_module(DRIVER_TYPES[driver_type]), driver_type)
+
+
+def __getattr__(name: str) -> Any:
+    """Resolve overlay classes lazily so importing this package does not require pynq."""
+    if name in DRIVER_TYPES:
+        return get_driver_class(name)
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
