+++ driver/finn_plus_driver/__main__.py
@@ -1,0 +1,6 @@
+"""Allow running the driver CLI via ``python -m finn_plus_driver``."""
+
+from finn_plus_driver.cli import driver_cli
+
+if __name__ == "__main__":
+    driver_cli()
