+++ driver/finn_plus_driver/overlays/__init__.py
@@ -1,0 +1,13 @@
+"""Overlay classes for FINN-generated accelerators."""
+
+from finn_plus_driver.overlays.dma import FINNDMAOverlay
+from finn_plus_driver.overlays.dma_instrumentation import FINNDMAInstrumentationOverlay
+from finn_plus_driver.overlays.instrumentation import FINNInstrumentationOverlay
+from finn_plus_driver.overlays.live_fifo import FINNLiveFIFOOverlay
+
+__all__ = [
+    "FINNDMAInstrumentationOverlay",
+    "FINNDMAOverlay",
+    "FINNInstrumentationOverlay",
+    "FINNLiveFIFOOverlay",
+]
