+++ driver/finn_plus_driver/packing.py
@@ -27,17 +27,18 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-"""Helpers for packing and unpacking FINN tensor data representations.
+"""Packing and unpacking of FINN tensor data representations.
 
-The runtime-relevant subset of this module is mirrored in ``finn_plus_driver.packing`` of
-the separately published ``finn-plus-driver`` package, so that PYNQ boards do not have to
-install the full compiler. Keep both sides in sync when changing either.
+This module is the runtime-only subset of ``finn.util.data_packing`` from the main
+finn-plus package. Only the functions needed to move data between numpy arrays and the
+bit-level representation expected by a generated accelerator are included here, so that a
+PYNQ board does not have to install the full compiler (onnx, onnxruntime, brevitas, ...).
+
+Keep the implementations in sync with ``src/finn/util/data_packing.py`` in
+https://github.com/eki-project/finn-plus when changing either side.
 """
 
-import binascii
 import numpy as np
-import os
-import sys
 from bitstring import BitArray
 from qonnx.core.datatype import DataType
 from qonnx.util.basic import roundup_to_integer_multiple
@@ -118,15 +119,6 @@ def hexstring2npbytearray(hexstring, remove_prefix="0x"):
     return np.asarray(bytearray.fromhex(hexstring), dtype=np.uint8)
 
 
-def npbytearray2hexstring(npbytearray, prefix="0x"):
-    """Convert a NumPy array of uint8 dtype into a hex string.
-
-    Example:
-    npbytearray2hexstring(array([15,  1], dtype=uint8)) = "0x0f01"
-    """
-    return prefix + binascii.hexlify(bytearray(npbytearray)).decode("utf-8")
-
-
 def pack_innermost_dim_as_hex_string(
     ndarray, dtype, pad_to_nbits, reverse_inner=False, prefix="0x"
 ):
@@ -155,183 +147,6 @@ def pack_innermost_dim_as_hex_string(
         return array2hexstring(x, dtype, pad_to_nbits, reverse=reverse_inner, prefix=prefix)
 
     return np.apply_along_axis(fun, ndarray.ndim - 1, ndarray)
-
-
-def unpack_innermost_dim_from_hex_string(
-    ndarray, dtype, out_shape, packedBits, reverse_inner=False
-):
-    """Convert a NumPy array of hex strings into a FINN NumPy array by unpacking
-    the hex strings into the specified data type. out_shape can be specified
-    such that any padding in the packing dimension is removed. If reverse_inner
-    is set, the innermost unpacked dimension will be reversed."""
-    if type(ndarray) is not np.ndarray:
-        raise Exception(
-            """unpack_innermost_dim_from_hex_string needs ndarray
-        as input"""
-        )
-    if ndarray.dtype.kind not in {"U", "S"}:
-        raise Exception(
-            """unpack_innermost_dim_from_hex_string needs ndarray of
-        hex strings as input"""
-        )
-    # convert ndarray into flattened list
-    data = ndarray.flatten().tolist()
-    targetBits = dtype.bitwidth()
-    # calculate outer and inner dim shapes
-    outer_dim_elems = 1
-    for dim in range(len(out_shape) - 1):
-        outer_dim_elems = outer_dim_elems * out_shape[dim]
-    inner_dim_elems = out_shape[-1]
-
-    array = []
-    if dtype.is_fixed_point():
-        # convert fixed point as signed integer
-        conv_dtype = DataType["INT" + str(targetBits)]
-    else:
-        conv_dtype = dtype
-    for outer_elem in range(outer_dim_elems):
-        ar_list = []
-        ar_elem = data[0]
-        data.pop(0)
-        ar_elem = ar_elem.split("x")
-        ar_elem_bin = bin(int(ar_elem[1], 16))[2:].zfill(packedBits)
-        ar_elem_bin = [int(x) for x in ar_elem_bin]
-
-        ar_elem_bin.reverse()
-        for i in range(inner_dim_elems):
-            upper_limit = (i + 1) * targetBits
-            lower_limit = i * targetBits
-            elem = ar_elem_bin[lower_limit:upper_limit]
-            elem.reverse()
-            elem_str = "".join(map(str, elem))
-            if conv_dtype == DataType["FLOAT16"]:
-                ar_list.append(np.float16(BitArray(bin=elem_str).float16))
-            elif conv_dtype == DataType["FLOAT32"]:
-                ar_list.append(BitArray(bin=elem_str).float)
-            elif conv_dtype.is_integer():
-                ar_list.append(int(elem_str, 2))
-            else:
-                raise Exception("Not implemented for conv_dtype " + conv_dtype.name)
-        # reverse inner dimension back to "normal" positions
-        if reverse_inner is False:
-            ar_list.reverse()
-
-        # interpret output values correctly
-
-        # interpret values as bipolar
-        if conv_dtype == DataType["BIPOLAR"]:
-            ar_list = [2 * x - 1 for x in ar_list]
-        # interpret values as signed values
-        elif conv_dtype.signed() and conv_dtype.is_integer():
-            mask = 2 ** (conv_dtype.bitwidth() - 1)
-            ar_list = [-(x & mask) + (x & ~mask) for x in ar_list]
-
-        array.append(ar_list)
-    npy_dtype = np.float16 if conv_dtype == DataType["FLOAT16"] else np.float32
-    array = np.asarray(array, dtype=npy_dtype).reshape(out_shape)
-    if dtype.is_fixed_point():
-        # convert signed integer to fixed point by applying scale
-        array = array * dtype.scale_factor()
-    return array
-
-
-def numpy_to_hls_code(ndarray, dtype, hls_var_name, pack_innermost_dim=True, no_decl=False):
-    """Return C++ code representation of a numpy ndarray with FINN DataType
-    dtype, using hls_var_name as the resulting C++ variable name. If
-    pack_innermost_dim is specified, the innermost dimension of the ndarray
-    will be packed into a hex string using array2hexstring. If no_decl is
-    set to True, no variable name and type will be generated as part of the
-    emitted string.
-    """
-    hls_dtype = dtype.get_hls_datatype_str()
-    if type(ndarray) is not np.ndarray or ndarray.dtype != np.float32:
-        # try to convert to a float numpy array (container dtype is float)
-        ndarray = np.asarray(ndarray, dtype=np.float32)
-    if pack_innermost_dim:
-        idimlen = ndarray.shape[-1]
-        idimbits = idimlen * dtype.bitwidth()
-        idimbits = roundup_to_integer_multiple(idimbits, 4)
-        ndarray = pack_innermost_dim_as_hex_string(ndarray, dtype, idimbits)
-        hls_dtype = "ap_uint<%d>" % idimbits
-    ndims = ndarray.ndim
-    # add type string and variable name
-    # e.g. "const ap_uint<64>" "weightMem0"
-    ret = "%s %s" % (hls_dtype, hls_var_name)
-    # add dimensions
-    for d in range(ndims):
-        ret += "[%d]" % ndarray.shape[d]
-    orig_printops = np.get_printoptions()
-    np.set_printoptions(threshold=sys.maxsize)
-
-    # define a function to convert a single element into a C++ init string
-    # a single element can be a hex string if we are using packing
-    def elem2str(x):
-        """Format a single element for C++ array initialization."""
-        if type(x) is str or type(x) is np.str_:
-            return '%s("%s", 16)' % (hls_dtype, x)
-        if type(x) is np.float32:
-            if dtype.is_integer():
-                return str(int(x))
-            return str(x)
-        raise Exception("Unsupported type for numpy_to_hls_code")
-
-    strarr = np.array2string(ndarray, separator=", ", formatter={"all": elem2str})
-    np.set_printoptions(**orig_printops)
-    strarr = strarr.replace("[", "{").replace("]", "}")
-    if no_decl:
-        ret = strarr + ";"
-    else:
-        ret = ret + " = \n" + strarr + ";"
-    return ret
-
-
-def npy_to_rtlsim_input(input_file, input_dtype, pad_to_nbits, reverse_inner=True):
-    """Convert the multidimensional NumPy array of integers (stored as floats)
-    from input_file into a flattened sequence of Python arbitrary-precision
-    integers, packing the innermost dimension. See
-    finn.util.basic.pack_innermost_dim_as_hex_string() for more info on how the
-    packing works. If reverse_inner is set, the innermost dimension will be
-    reversed prior to packing."""
-    pad_to_nbits = roundup_to_integer_multiple(pad_to_nbits, 4)
-    if issubclass(type(input_file), np.ndarray):
-        inp = input_file
-    elif os.path.isfile(input_file):
-        inp = np.load(input_file)
-    else:
-        raise Exception("input_file must be ndarray or filename for .npy")
-    if (
-        inp.shape[-1] == 1
-        and input_dtype.is_integer()
-        and input_dtype.get_canonical_name() != "BIPOLAR"
-    ):
-        mask = (1 << input_dtype.bitwidth()) - 1
-        packed_data = inp.flatten().astype(input_dtype.to_numpy_dt())
-        packed_data = [int(x) & mask for x in packed_data]
-    else:
-        packed_data = pack_innermost_dim_as_hex_string(
-            inp, input_dtype, pad_to_nbits, reverse_inner=reverse_inner
-        )
-        packed_data = packed_data.flatten()
-        packed_data = [int(x[2:], 16) for x in packed_data]
-    return packed_data
-
-
-def rtlsim_output_to_npy(output, path, dtype, shape, packedBits, targetBits, reverse_inner=True):
-    """Convert a flattened sequence of Python arbitrary-precision integers
-    output into a NumPy array, saved as npy file at path. Each arbitrary-precision
-    integer is assumed to be a packed array of targetBits-bit elements, which
-    will be unpacked as the innermost dimension of the NumPy array. If path is
-    not None it will also be saved as a npy file."""
-    # TODO should have its own testbench?
-    output = np.asarray([hex(int(x)) for x in output])
-    out_array = unpack_innermost_dim_from_hex_string(
-        output, dtype, shape, packedBits=packedBits, reverse_inner=reverse_inner
-    )
-    # make copy before saving the array
-    out_array = out_array.copy()
-    if path is not None:
-        np.save(path, out_array)
-    return out_array
 
 
 def finnpy_to_packed_bytearray(
@@ -583,17 +398,3 @@ def packed_bytearray_to_finnpy_float(
     if reverse_inner:
         unpacked_float = np.flip(unpacked_float, -1)
     return unpacked_float
-
-
-def to_external_tensor(init, w_dtype):
-    """Return an appropriately formatted and packed numpy byte array for given
-    external parameter tensor."""
-    weight_width = init.shape[1] * w_dtype.bitwidth()
-    weight_width_padded = roundup_to_integer_multiple(weight_width, 4)
-    hex_init = pack_innermost_dim_as_hex_string(init, w_dtype, weight_width_padded, prefix="0x")
-    ext_weight = np.array([], dtype=np.uint8)
-    for line in hex_init:
-        array_line = [x for x in reversed(hexstring2npbytearray(line, remove_prefix="0x"))]
-        ext_weight = np.append(ext_weight, array_line)
-
-    return ext_weight
