+++ driver/pyproject.toml
@@ -1,0 +1,50 @@
+[project]
+name = "finn-plus-driver"
+dynamic = ["version"]
+description = "PYNQ runtime driver for FPGA accelerators generated by finn-plus"
+requires-python = ">=3.10"
+maintainers = [
+    { name = "Felix Jentzsch", email = "felix.jentzsch@upb.de" },
+    { name = "Christoph Berganski", email = "christoph.berganski@upb.de" },
+    { name = "Linus Jungemann", email = "linus.jungemann@upb.de" },
+    { name = "Bjarne Wintermann", email = "bjarne.wintermann@upb.de" },
+]
+
+[tool.poetry.requires-plugins]
+poetry-dynamic-versioning = { version = ">=1.0.0,<2.0.0", extras = ["plugin"] }
+
+[tool.poetry-dynamic-versioning]
+enable = true
+
+[project.urls]
+"GitHub" = "https://github.com/eki-project/finn-plus"
+"EKI Project" = "https://www.eki-project.tech/"
+
+[tool.poetry]
+name = "finn-plus-driver"
+version = "0.0.0"
+description = "PYNQ runtime driver for FPGA accelerators generated by finn-plus"
+readme = "README.md"
+packages = [{ include = "finn_plus_driver" }]
+
+# NOTE: pynq is intentionally not declared as a dependency. It is part of the
+# PYNQ board image and cannot be installed from PyPI in a meaningful way.
+# Version constraints are kept loose on purpose so that installing this package
+# on a board does not force an upgrade of packages the PYNQ image depends on.
+[tool.poetry.dependencies]
+python = ">=3.10"
+numpy = ">=1.21"
+bitstring = ">=4.0"
+qonnx = ">=1.0.0"
+click = ">=8.0"
+matplotlib = ">=3.4"
+h5py = ">=3.0"
+pillow = ">=9.0"
+finn-dataset-loading = ">=0.0.5"
+
+[project.scripts]
+finn-plus-driver = "finn_plus_driver.cli:driver_cli"
+
+[build-system]
+requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
+build-backend = "poetry_dynamic_versioning.backend"
